Some ui changes and pose thumbnail stuff
