+++ Resources/Materials/PumkinsThumbnailCamUnlit.mat
