+++ Resources/Presets/Cameras/Green Centered.asset
