+++ Resources/Presets/Cameras/Pink.asset
