+++ Resources/Presets/Cameras/eye.asset
