+++ Resources/Presets/Poses/hands behind.asset
@@ -110,3 +110,5 @@ MonoBehaviour:
   - 0.8162024
   - 0.81042
   presetMode: 0
+  transformPaths: []
+  transformRotations: []
