+++ Resources/Presets/Poses/standing.asset
